+++ packages/@sparrow-workspaces/src/features/ai-request-explorer/layout/AiRequestExplorer.svelte
@@ -4,11 +4,10 @@
     ModelSector,
     RequestNavigator,
     RequestAuth,
-    RequestName,
     ChatBot,
     RequestDoc,
     AiConfigs,
-    ConversationHistoryItem,
+    GetCode,
   } from "../components";
   import { Splitpanes, Pane } from "svelte-splitpanes";
   import type {
@@ -29,27 +28,10 @@
   import { onDestroy, onMount } from "svelte";
   import { writable } from "svelte/store";
   import { disabledModelFeatures, modelCodeTemplates } from "../constants";
-
-  import {
-    BotRegular,
-    SettingsRegular,
-    BotSparkleRegular,
-    DismissRegular,
-    CopyIcon,
-    CopyIcon2,
-    TickIcon,
-    CopyRegular,
-  } from "@sparrow/library/icons";
-  import { Button, Modal, notifications } from "@sparrow/library/ui";
+  import { SettingsRegular, BotSparkleRegular } from "@sparrow/library/icons";
+  import { Modal } from "@sparrow/library/ui";
   import { SaveAsCollectionItem } from "../../save-as-request";
   import { TabTypeEnum } from "@sparrow/common/types/workspace/tab";
-  import { Editor, Select } from "@sparrow/library/forms";
-  import type { CodeTemplateLanguage } from "../types";
-  import {
-    CodeEditorLanguageType,
-    CodeTemplateLanguageType,
-  } from "../types/ai-request";
-  import { Sleep } from "@sparrow/common/utils";
 
   export let tab: Observable<Tab>;
   export let collections: Observable<CollectionDocument[]>;
@@ -178,11 +160,6 @@
     isConversationHistoryPanelOpened = false;
   };
 
-  // $: {
-  //   if ($tab?.property?.aiRequest)
-  //     console.log("tab :>> ", $tab?.property?.aiRequest);
-  // }
-
   const handleOnClickUpdateRequestAuth = async () => {
     if (isConversationHistoryPanelOpened) {
       isConversationHistoryLoading = true;
@@ -193,52 +170,9 @@
     onUpdateRequestAuth();
   };
 
-  let codeTemplateLanguage: CodeEditorLanguageType =
-    CodeEditorLanguageType.nodejs;
-  let selectedCodeTemplateId = CodeTemplateLanguageType.NODEJS;
-  let codeTemplate: string;
-  let isCodeTemplateBeautified = true;
-
-  const handleCodeTemplateChange = (
-    codeTemplateId: CodeTemplateLanguageType,
-  ) => {
-    const provider = $tab?.property?.aiRequest?.aiModelProvider;
-    const variant = $tab?.property?.aiRequest?.aiModelVariant;
-    const apiKey = $tab?.property?.aiRequest?.auth?.apiKey?.authValue;
-    const configuration = $tab?.property?.aiRequest?.configurations;
-
-    selectedCodeTemplateId = codeTemplateId;
-    codeTemplateLanguage = CodeEditorLanguageType[selectedCodeTemplateId];
-
-    // Get the template function
-    const templateFunction =
-      modelCodeTemplates[provider][selectedCodeTemplateId]["code_template"];
-
-    // Call the template function with parameters to get the actual code string
-    codeTemplate = templateFunction(variant, apiKey, configuration[provider]);
-    updateCodeTemplateBeautifiedState(true);
-  };
-
   let isGetCodePopupOpen = false;
-  let isCodeCopied = false;
   const onClickOpenGetCodePopup = async () => {
-    const provider = $tab?.property?.aiRequest?.aiModelProvider;
-    const variant = $tab?.property?.aiRequest?.aiModelVariant;
-    const apiKey = $tab?.property?.aiRequest?.auth?.apiKey?.authValue;
-    const configuration = $tab?.property?.aiRequest?.configurations;
-
-    codeTemplateLanguage = CodeEditorLanguageType.nodejs;
-    selectedCodeTemplateId = CodeTemplateLanguageType.NODEJS;
-    const templateFunction =
-      modelCodeTemplates[provider][selectedCodeTemplateId]["code_template"];
-    codeTemplate = templateFunction(variant, apiKey, configuration[provider]);
-
     isGetCodePopupOpen = true;
-    updateCodeTemplateBeautifiedState(true);
-  };
-
-  const updateCodeTemplateBeautifiedState = (value: boolean) => {
-    isCodeTemplateBeautified = value;
   };
 </script>
 
@@ -458,89 +392,12 @@
       isGetCodePopupOpen = false;
     }}
   >
-    <div
-      class="d-flex align-items-center justify-content-between mt-2 mb-2 w-100 gap-2"
-    >
-      <div class="flex-grow-1" style="max-width: 75%;">
-        <Select
-          id={"code-selector"}
-          data={[
-            {
-              id: CodeTemplateLanguageType.NODEJS,
-              name: "NodeJs",
-              disabled: false,
-              hide: false,
-            },
-            {
-              id: CodeTemplateLanguageType.JSON,
-              name: CodeEditorLanguageType.json,
-              disabled: false,
-              hide: false,
-            },
-            {
-              id: CodeTemplateLanguageType.PYTHON,
-              name: CodeEditorLanguageType.python,
-              disabled: false,
-              hide: false,
-            },
-            {
-              id: CodeTemplateLanguageType.CURL,
-              name: CodeEditorLanguageType.curl,
-              disabled: false,
-              hide: false,
-            },
-          ]}
-          titleId={selectedCodeTemplateId}
-          onclick={(id) => {
-            handleCodeTemplateChange(id);
-          }}
-          menuItem={"v2"}
-          minHeaderWidth={"100%"}
-          maxHeaderWidth={"100%"}
-          minBodyWidth={"100%"}
-          variant={"light-violet"}
-          position={"absolute"}
-          zIndex={200}
-        />
-      </div>
-
-      <div class="flex-shrink-0">
-        <Button
-          title={isCodeCopied ? "Copied" : "Copy Code"}
-          size={"small"}
-          type={"teritiary-regular"}
-          startIcon={isCodeCopied ? TickIcon : CopyRegular}
-          iconSize={14}
-          onClick={async () => {
-            if (isCodeCopied) return;
-            try {
-              await navigator.clipboard.writeText(codeTemplate);
-              isCodeCopied = true;
-              notifications.success("Code copied to clipboard.");
-              await new Sleep().setTime(1500).exec();
-              isCodeCopied = false;
-            } catch (err) {
-              notifications.success("Failed to copy the code.");
-            }
-          }}
-        />
-      </div>
-    </div>
-
-    <div
-      class="request-body position-relative w-100"
-      style="background-color: var(--bg-ds-neutral-900); border-radius: 2px; height: 260px; border: 1px solid var(--border-color, rgba(255,255,255,0.1));"
-    >
-      <Editor
-        bind:lang={codeTemplateLanguage}
-        bind:value={codeTemplate}
-        on:change={() => {}}
-        isEditable={false}
-        autofocus={false}
-        isBodyBeautified={isCodeTemplateBeautified}
-        beautifySyntaxCallback={updateCodeTemplateBeautifiedState}
-      />
-    </div>
+    <GetCode
+      selectedModelVariant={$tab?.property?.aiRequest?.aiModelVariant}
+      aiModelProvider={$tab?.property?.aiRequest?.aiModelProvider}
+      providerApiKey={$tab?.property?.aiRequest?.auth?.apiKey?.authValue}
+      configurations={$tab?.property?.aiRequest?.configurations}
+    />
   </Modal>
 {/if}
