+++ packages/@sparrow-workspaces/src/features/ai-request-explorer/types/index.ts
@@ -1,1 +1,1 @@
-export { type CodeTemplateLanguage } from "./ai-request"
+export { type CodeTemplateLanguage, CodeTemplateLanguageType, CodeEditorLanguageType } from "./ai-request"
